# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for EVE Online trading, manufacturing and profit optimization\n\nFeatures:\n- Intra-region trading route calculation\n- Character skills and ship fitting integration\n- Deterministic cargo/warp speed calculations\n- Market data caching with staleness tracking\n- EVE SSO authentication",
        "title": "EVE-O-Provit API",
        "contact": {
            "name": "EVE-O-Provit Team",
            "url": "https://github.com/Sternrassler/eve-o-provit",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/items/search": {
            "get": {
                "description": "Search for EVE Online items by name (fuzzy matching)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trading"
                ],
                "summary": "Search EVE items",
                "parameters": [
                    {
                        "minLength": 3,
                        "type": "string",
                        "description": "Search query (min 3 characters)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Maximum results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ItemSearchResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/trading/routes/calculate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculate optimal intra-region trading routes with profit analysis\nUses character skills and ship fitting for accurate cargo capacity\nSupports volume filtering for liquidity-based selection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trading"
                ],
                "summary": "Calculate trading routes",
                "parameters": [
                    {
                        "description": "Route calculation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TradingRouteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TradingRouteResponse"
                            }
                        }
                    },
                    "206": {
                        "description": "Partial results (timeout)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TradingRouteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "Detailed error message"
                }
            }
        },
...
        "ItemSearchResult": {
            "type": "object",
            "properties": {
                "group_name": {
                    "type": "string",
                    "example": "Mineral"
                },
                "name": {
                    "type": "string",
                    "example": "Tritanium"
                },
                "type_id": {
                    "type": "integer",
                    "example": 34
                }
            }
        },
...
        "TradingRouteRequest": {
            "type": "object",
            "required": [
                "region_id",
                "type_ids"
            ],
            "properties": {
                "cargo_capacity": {
                    "type": "number",
                    "example": 9656.9
                },
                "character_id": {
                    "type": "integer",
                    "example": 12345678
                },
                "max_investment": {
                    "type": "number",
                    "example": 1000000000
                },
                "max_jumps": {
                    "type": "integer",
                    "example": 5
                },
                "region_id": {
                    "type": "integer",
                    "example": 10000002
                },
                "ship_type_id": {
                    "type": "integer",
                    "example": 650
                },
                "type_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        34,
                        35,
                        36
                    ]
                }
            }
        },
        "TradingRouteResponse": {
            "type": "object",
            "properties": {
                "buy_location": {
                    "type": "string",
                    "example": "Jita IV - Moon 4 - Caldari Navy Assembly Plant"
                },
                "buy_price": {
                    "type": "number",
                    "example": 5.5
                },
                "investment": {
                    "type": "number",
                    "example": 550000
                },
                "jumps": {
                    "type": "integer",
                    "example": 5
                },
                "profit": {
                    "type": "number",
                    "example": 50000
                },
                "profit_per_jump": {
                    "type": "number",
                    "example": 10000
                },
                "quantity": {
                    "type": "integer",
                    "example": 100000
                },
                "revenue": {
                    "type": "number",
                    "example": 600000
                },
                "roi_percent": {
                    "type": "number",
                    "example": 9.09
                },
                "sell_location": {
                    "type": "string",
                    "example": "Amarr VIII (Oris) - Emperor Family Academy"
                },
                "sell_price": {
                    "type": "number",
                    "example": 6
                },
                "type_id": {
                    "type": "integer",
                    "example": 34
                },
                "type_name": {
                    "type": "string",
                    "example": "Tritanium"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "EVE SSO Bearer Token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}